added front end for view post and edit post, event listeners and routes for navigation
