+++ public/js/index.js
@@ -1,6 +1,3 @@
-//need event listeners to navigate pages
-//event listener for home, login, dashboard
-//event listener to view a specific post and comments
 const homeButtonEl = document.querySelector("#home");
 const loginButtonEl = document.querySelector("#login");
 const dashboardButtonEl = document.querySelector("#dashboard");
@@ -9,7 +6,6 @@ const loginFooter = document.querySelector("#login-footer");
 const dashboardFooter = document.querySelector("#dashboard-footer");
 
 const createButtonEl = document.querySelector("#create");
-const editPostButtonEl = document.querySelector("#edit");
 
 const displayHomepage = async (event) => {
   event.preventDefault();
@@ -66,20 +62,6 @@ const displayCreatePage = async (event) => {
   }
 };
 
-const displayEditPage = async (event) => {
-  event.preventDefault();
-
-  const response = await fetch("/update", {
-    method: "GET",
-  });
-  console.log(response);
-  if (response.ok) {
-    document.location.replace("/update");
-  } else {
-    alert("Failed to load Update page");
-  }
-};
-
 //need event listener for submit button on login page
 homeButtonEl.addEventListener("click", displayHomepage);
 homeFooter.addEventListener("click", displayHomepage);
@@ -89,4 +71,3 @@ dashboardButtonEl.addEventListener("click", displayDashboard);
 dashboardFooter.addEventListener("click", displayDashboard);
 
 createButtonEl.addEventListener("click", displayCreatePage);
-editPostButtonEl.addEventListener("click", displayEditPage);
