+++ views/dashboard.handlebars
@@ -9,10 +9,12 @@
             <h5 class="post-title">{{post.title}}</h5>
             <p class="post-author-id">Author: {{post.author_username}}</p>
             <p class="post-content">{{post.content}}</p>
-            <a href="#" class="btn btn-primary">View Comments</a>
+            <a href="#" id="view-post" class="btn btn-primary">View Post</a>
           </div>
         </div>
       </div>
     {{/each}}
+    <button class="button" id="view-post">View Post</button>
   </div>
 </div>
+<script src="/js/dashBoard.js"></script>
