+++ views/layouts/main.handlebars
@@ -56,9 +56,7 @@
                 <li class="nav-item">
                   <a class="nav-link" id="create" href="#">Create A Post</a>
                 </li>
-                <li class="nav-item">
-                  <a class="nav-link" id="edit" href="#">Update A Post</a>
-                </li>
+                
               </ul>
             </form>
           </div>
