+++ views/updatepost.handlebars
@@ -1,18 +1,33 @@
 <div class="col-md-4 mb-4">
-    <div class="card">
-        <div class="card-body">
-            <h5 class="card-title">Update Your Post</h5>
-            <form action="process_post.php" method="post">
-                <div class="mb-3">
-                    <label for="postTitle" class="form-label">Title:</label>
-                    <input type="text" class="form-control" id="postTitle" name="postTitle" value={{post.title}}>
-                </div>
-                <div class="mb-3">
-                    <label for="postContent" class="form-label">Content:</label>
-                    <textarea class="form-control" id="postContent" name="postContent" rows="3" value={{post.content}}></textarea>
-                </div>
-                <button type="submit" class="btn btn-primary">update a post</button>
-            </form>
+  <div class="card">
+    <div class="card-body">
+      <h5 class="card-title">Post Title</h5>
+      <h6 class="">Created on ...</h6>
+      <form action="process_post.php" method="post">
+        <div class="mb-3">
+          <label for="postTitle" class="form-label">Title:</label>
+          <input
+            type="text"
+            class="form-control"
+            id="postTitle"
+            name="postTitle"
+            value={{post.title}}
+          />
         </div>
+        <div class="mb-3">
+          <label for="postContent" class="form-label">Content:</label>
+          <textarea
+            class="form-control"
+            id="postContent"
+            name="postContent"
+            rows="3"
+            value={{post.content}}
+          >To be populated with the post.content</textarea>
+        </div>
+        <button type="submit" class="btn btn-primary">update a post!</button>
+        <a href="#" id="go-back" class="btn btn-primary">Back to Post</a>
+      </form>
     </div>
+  </div>
 </div>
+<script src="/js/updatePost.js"></script>
