+++ views/viewpost.handlebars
@@ -1,0 +1,37 @@
+<div class="container text-center">
+  <div class="row">
+    <div class="col">
+      <h1 class="text">
+        Card Title
+      </h1>
+      <h3>created on...</h3>
+
+      <p class="card">
+        Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod
+        tempor incididunt ut labore et dolore magna aliqua. Purus in massa
+        tempor nec feugiat nisl pretium fusce id. Enim facilisis gravida neque
+        convallis a cras semper auctor. Pellentesque massa placerat duis
+        ultricies. Elit pellentesque habitant morbi tristique senectus et netus.
+        Diam vulputate ut pharetra sit amet. Duis at consectetur lorem donec
+        massa sapien faucibus. Vitae nunc sed velit dignissim sodales ut. Nunc
+        non blandit massa enim nec dui. Venenatis urna cursus eget nunc
+        scelerisque viverra mauris.s</p>
+      <div class="card-body">
+        <a href="#" id="edit-post" class="btn btn-primary">Edit post</a>
+        <a href="#" id="go-back" class="btn btn-primary">Back to dashboard</a>
+      </div>
+
+      <div class="card">
+        <div class="card-body">
+          <h5 class="post-title">comments </h5>
+          <p>comment1</p>
+          <p>comment1</p>
+          <p>comment1</p>
+
+        </div>
+      </div>
+    </div>
+  </div>
+</div>
+
+<script src="/js/viewPost.js"></script>
